- Fix dynamic mask not being flagged for duplicate objects - Allow a negative sign to be placed in front of Relevant object names.  For example in face matching, if you use "person, car, badguyname, -GOODGUYNAME", then it would be triggered for badguyname but not goodguyname - Separate sighthound license plate detection rather than merging - Further optimizations of how linked and refinement servers combine results

+++ src/UI/ClsSightHound.cs
@@ -144,7 +144,7 @@ namespace AITool
         public string Name { get; set; }
 
         [JsonProperty("confidence")]
-        public double Confidence { get; set; }
+        public float Confidence { get; set; }
     }
 
     public class SightHoundCharacter
